添加 NSCT 图像重建测试覆盖文档，包含形状一致性、数值精度、一对一一致性、参数化和边界情况测试

+++ scripts/benchmark_nsct_backends.py
@@ -60,7 +60,7 @@ def parse_args() -> argparse.Namespace:
     parser.add_argument(
         "--dtype",
         choices=("float32", "float64"),
-        default="float32",
+        default="float64",
         help="Floating point precision to use for both backends (default: %(default)s).",
     )
     parser.add_argument(
